+++ src/molpacd/__init__.py
@@ -2,18 +2,31 @@ from __future__ import annotations
 
 from molpacd._version import __version__
 from molpacd.capper import add_caps, analyze_structure, remove_caps
+from molpacd.cavity_lipids import remove_cavity_lipids
+from molpacd.cavity_lipids_pdb import remove_cavity_lipids_pdb
 from molpacd.io import read_structure, write_structure
-from molpacd.models import AnalysisResult, CapOptions, CapResult, RemoveResult
+from molpacd.models import (
+    AnalysisResult,
+    CapOptions,
+    CapResult,
+    CavityLipidOptions,
+    CavityLipidResult,
+    RemoveResult,
+)
 
 __all__ = [
     "AnalysisResult",
     "CapOptions",
     "CapResult",
+    "CavityLipidOptions",
+    "CavityLipidResult",
     "RemoveResult",
     "__version__",
     "add_caps",
     "analyze_structure",
     "read_structure",
     "remove_caps",
+    "remove_cavity_lipids",
+    "remove_cavity_lipids_pdb",
     "write_structure",
 ]